Added a new header for "internal" depsgraph API functions
These are mainly functions which can be used internally within the
Depsgraph, and/or are exported between various parts of the internals
but which shouldn't really be exposed to the outside world (i.e. other
parts of Blender outside of the depsgraph module).

Will review later whether this file is really needed or whether it can
be folded into some of the other headers)