Change argument types for low-level node management functions
Apparently it is not possible to simply pass on the list of arbitrary
arguments down the chain that easily (where the list grows and shrinks
according to the needs of the specific API used). So instead, always use
the full version.

Alternative here is to just have separate copies for each case?