Start of revamped design (#2) - New node type enum defines
This commit marks the start of the second attempt at building a new
depsgraph system for Blender. It still uses the base depsgraph types
from the first attempt, but this time, the node types we've chosen to
sit on top of these (for representing the necessary semantic concepts)
have been chosen to be more in line with the requirements we face in
terms of how exactly we're going to be able to use this for evaluating
stuff.