Change arguments to DEG_get/find/add_node() methods to use string/name arg to distinguish items
Previously, these used RNA Pointers, but that approach proves to be
useless in most cases we have (type already covers the basic honing in,
but then the pointers prove to be too coarse to do anything about callbacks).

So, now use the identifiers stored in nodes to perform this role.