Added "validate_links()" callback to node typeinfo
This callback is used to ensure that all "builtin" operations and/or
relationships between components used to enforce certain valid-evaluation-
order constraints are satisfied. Having some of these in graph building
would be a bit too much of an extra burden to have, when much of these
are pretty much standardised actions that must take place.