Support for debug graphviz output during the build process.
This uses two callbacks which can be set prior to rebuilding the
depsgraph. These will be called when nodes or relations are added, which
leaves the actual implementation up to the caller with minimal impact
on the depsgraph code.

These functions are stubs in non-debug modes, so they don't cause
overhead.

The Scene.depsgraph_rebuild() function implements this via an optional
filename argument. If this is specified it will dump build steps into a
sequence of files using _<N> suffixes with a running number N.