Basic callback in ModifierTypeInfo for registering dependencies in the new depsgraph.
For now this lives alongside the old callback until porting is complete
and the old callback can be removed.

The modifier cb uses a DepsNodeHandle symbolic type for representing the
current modifier node itself, so it doesn't get into conflict with the
actual DepsNode class. This will also incorporate a pointer to Depsgraph
and possibly other state info internally, so only a single handle
argument is needed.