Moved DEG_add_new_relation into Depsgraph.
Note: This function doesn't take a Depsgraph argument. Currently the
ownership for relation instances is shared between the nodes being
connected. This is not ideal, it prevents good memory management and
leads to ugly corner cases when freeing relations. Eventually relations
should be allocated and stored by the Depsgraph, since they only live
inside the depsgraph data anyway (ownership is with the graph).