Operator for copying hair data from old particle systems to the new hair data structure.
This will allow a quick initialization of the hair data with real world
examples, so we can test the hair simulation etc. without having to
write setup code or fully fledged tools first.

Possibly could also allow porting of old blend file data, but that is
of secondary importance for now.