Implemented hair frame calculation using parallel transport along curve segments.
This can be done on-the-fly but rotating the root hair frame
incrementally according to each hair segment angle. Note: a similar
approach can be found in the Blender curve code for calculating twist
vectors.

A novel feature of the hair frame calculation is the use of smoothed
hair curves. This prevents artifacts in case of sharp segment angles by
removing small-scale noise from the curve.