Base support for hair root animation.
Note: This does not work currently, because of the way the solver is
reconstructed in every time step currently, and so does not provide
actual changes in the transform.

Note2: The mesh sampling system does not support changes in DerivedMesh
topology so far. It can be necessary to move the hair system to the
bottom of the modifier stack, because the hair roots are created for
the final mesh version and for animation support it can only use the
mesh being passed to the modifier. Any later subsurf etc. (constructive
modifier) will alter the topology and cause a difference in the DM used
for hair roots vs. DM used for calculating deformation.