Fix problem of hair root animation by keeping the solver around as a runtime instance.
The SolverData also functions as a short-term history for constructing
the hair root animation/deformation curves. This is necessary because
the DerivedMesh data on its own does not have interpolation intervals
nor velocity information.

It also means that we have to keep this data alive over multiple frames
in order to do root animation properly. So now the solver is created as
a runtime instance.