Integrate hair solver stepping with the global scene stepping used for Bullet physics.
This is a preparation to using Bullet for hair-mesh collisions. This
is much more accurate and requires less secondary interpolation when
the hair solver is in sync with the Bullet stepping.

Bullet uses essentially the same time step concepts as the hair solver
did, i.e. the main "step" (1 frame) is divided into fixed-size "ticks".
This is paramount for stability of the solver. The hair solver could
eventually use a multiple or integer fraction of the Bullet tick size
if more accuracy or performance is necessary.