More sensible defaults and ranges for the hair sim settings.
Note that this will probably change again in the future. The goal is to
make the simulation as stable as possible in the default ui range of
properties (artists can override soft limits, but then instability has
to be expected).

Currently spring forces and damping depend somewhat on the time step
size and hair segment length. Also reparameterizing properties a bit
could help to increase accuracy in artistically interesting ranges.