Simple hair contact point debugging using a ghost object per hair point.
The rbDynamicsWorld is now also exposed in the types header for the RBI
API, because currently we need direct access to some Bullet properties.

User pointers for collision objects now point to a generic
rbCollisionObject base class, so we don't run into trouble when
explicitly casting to rbRigidBody.

Points can not be copied explicitly any more, because it makes internal
point variables invalid that are hard to track. Bullet itself also does
not like direct copies of btRigidBody ...