Make Bullet world stepping and rebuild less dependent on point cache.
This means the Bullet rigid body world will be created even when a baked
rigid body cache exists and stepping will also happen regardless of
baked or cached results being available.

The rationale is that the Bullet DynamicsWorld stepping may be used for
other systems as well (hair, later particles) which do not necessarily
use the same cache system. Rather than disabling stepping and collision
detection altogether the cached RigidBodies should rather be set to
kinematic, so that other un-cached participants can still use them for
collisions. This will also be an important prerequisite for later
layering of simulations with pre-cached or animated elements.