Avoid adding all the hair point rigid bodies to the CollisionWorld.
Instead we now use a single box-shaped collision object for the whole
hair system, and then do direct contact tests for the point objects
when there is a potential contact. This avoids O(n^2) issues when
inserting lots of collision objects into a DbvtBroadphase.

The points now have a radius property of their own as well.