Implemented materials for hair systems.
This uses basically the same system as current particle hair: the hair
system is linked to the material via a slot index, which refers
to the object material slots. Unlike particle settings the hair system
is (for now) not a separate datablock, so the mapping is a bit more
consistent.