Modification to bending spring calculation to handle the initial direction of the hair curve.
Note that bending springs are a bit unstable for low bending stiffness
when there is not enough damping. This may be due to missing additional
derivatives on the previous segment (dependency of the frame), but
calculating these entirely could give a dense matrix instead ...

A more practical approach might be to make damping factors relative to
the stiffness and set a soft UI minimum of 10% or so. Values much below
that are not physically plausible anyway, given that any material has
internal friction on bending.