"Init" variants of shape constructor functions for simple implicit shapes.
These functions allow external memory allocation (using the associated
***_size values) for efficiency. An array or memory pool can then be
used to create lots of shapes and initialize them with the internal
Bullet data using the API.

Note that doing this for mesh-based shapes (triangle mesh, GImpact and
convex hull shapes) is not very useful, because they allocate data
internally and should not be created in massive numbers. The init
variants are supposed to be used for small fixed-size shapes (boxes,
spheres, etc.) which can then be combined into compound shapes.