Added API representation of manifold contact points and the direct contact test function from btCollisionWorld.
This function can be used for testing individual collision objects
(rigid bodies or ghosts) against objects in the world. It is only
useful for one-way interaction (rigid bodies don't react to the tested
object), but allows collision detection without adding objects into the
world. The current build process for the scene dynamics world is very
much centered on Blender Objects and extending it is difficult.