Added a second hair data layer to store modifier results.
This imitates the DerivedMesh principle from meshes. For hair the
modifiers will get the grooming or original growth result as input.
The result of the modifier eval will be written to the new data layer,
making a simple copy if modifiers are not used. Physics sim, rendering
and (depending of settings) drawing will then use the modified grooming
data.