Reset function for the final hair state before calculating hair dynamics and paths.
The integration with modifiers is still undefined: While it would be
nicer to have regular modifiers work on hair state as well, it is
probably not feasible with the current particle system, because all
these steps happen in a single update function. For integrating with
the modifier stack we would have to allow splitting this process into
individual parts, which is just not possible.