List of particle modifiers in particle systems and read/write code.
Note: Ideally modifiers for particles should be part of the
`ParticleSettings` datablock instead of `ParticleSystem`. The problem
with this is that many modifiers, as they are implemented for meshes,
will store specific data that is associated to the object instance
in the scene. Storing modifiers in the settings would make this
impossible and require some complicated syncing between particle systems
and their settings.